test(cli,rdf): pin query --entailment exit-3 boundary + dedup pack test fixture
Add a query_cli test asserting owl-direct/rif/d --entailment exit code 3
with a diagnostic, mirroring reason's unsupported-regime boundary test.
Extract the identical ~90-line build_fixture into crates/rdf/tests/common/mod.rs;
both pack tests now share it (no drift found).
